Refactor MaxPooling: Split into linear_block and max. Cast kernel to float32 for consistency.

+++ jacobinet/layers/pooling/average_pooling1d.py
@@ -54,7 +54,7 @@ class BackwardAveragePooling1D(BackwardLinearLayer):
             use_bias=False,
             trainable=False,
         )
-        kernel_ = np.ones(pool_size + [1, 1]) / np.prod(pool_size)
+        kernel_ = np.ones(pool_size + [1, 1], dtype='float32') / np.prod(pool_size)
         layer_t.kernel = keras.Variable(kernel_)
         layer_t.built = True
